# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "AuthRes": {
        "properties": {
          "token": {
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          }
        },
        "required": ["user", "token"],
        "type": "object"
      },
...
      "ErrorResponse": {
        "properties": {
          "error_msg": {
            "type": "string"
          }
        },
        "required": ["error_msg"],
        "type": "object"
      },
...
      "User": {
        "properties": {
          "_id": {
            "type": "string"
          },
          "_key": {
            "type": "string"
          },
          "_rev": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "first_name": {
            "type": "string"
          },
          "last_name": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "role": {
            "$ref": "#/components/schemas/Role"
          }
        },
        "required": [
          "_key",
          "_rev",
          "_id",
          "first_name",
          "last_name",
          "email",
          "password"
        ],
        "type": "object"
      }
    }
  },
  "info": {
    "contact": {
      "email": "andreapallotta.dev@gmail.com",
      "name": "Andrea Pallotta"
    },
    "description": "Secure REST API written in Rust & Axum",
    "license": {
      "name": ""
    },
    "title": "server",
    "version": "1.0.0"
  },
  "openapi": "3.0.3",
  "paths": {
...
    "/api/auth/refresh": {
      "get": {
        "operationId": "refresh",
        "parameters": [
          {
            "description": "user email",
            "in": "path",
            "name": "email",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthRes"
                }
              }
            },
            "description": "Return authenticated user"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Error generating jwt"
          }
        },
        "tags": ["server::requests::jwt"]
      }
    },
...
  },
...
}